Add another installed workload manifest detection/versioning
Currently the workload manifest json file does have a version field, but we're not incrementing it, and it sounds like it might change in the future.

The check tool was not checking for outdated manifest installs and so it wouldn't be checking for the latest packs referred to by those manifests.

This adds a method to look for the resolver provided installed manifest directories and manually inspects the nuspec file extracted from the installed manifest, to fetch its package id and package version which can then be used to check if the actual right version of the workload is installed.

This still keeps the check around to  load manifests from the resolver to ensure we have the workload id found by it too, but that check is not validating the workload manifest versions yet (again since we don't bump those yet anyway).