Wrap copy ops in su for non windows
It's been a lot of problems on mac with permissions on installed packs.  The summary seems to be that if extracting packages finally with the NuGet api's, but as super user, they may still retain their +x attribute, however the permissions are set to root and a normal user cannot execute.

The current workaround is to wrap any operation which needs to write a file or folder to a privileged location on mac in a method which first creates the dir/file in a temp location that is not privileged.  Finally, a shell command is run with sudo to copy the dir/file from the tmp location to the final privileged destination.

This also means we can delay asking for su until it's needed and not bug users about it when the tool is first run.