Delete things later
Deleting them when we wrap the call for a destination meant that for example, packs that would go in `packs/[pack-id]/[version]` would really have a destination of `packs` since the nuget packageextractor would figure out the `[pack-id]/[version]` part, so it would delete the root packs dir and everything in it.. oops!  This moves the delete call back to where we know the actual full install path and only delete it instead.