Look for templates installed in newest sdk folder
According to the docs, by default the newest sdk should be used by default in 'dotnet'.  Before this was checking in the sdk version folder which was the requested one from the manifest, but there could be a newer version installed on the system and that's where the templates would be installed, since they aren't really tied to an sdk version.

The best practice seems to just install to the latest and assume the latest, which makes sense since you probably don't have a global.json when invoking `dotnet new`.